[Revert] Revert custom methodsjs
